correctly configure network interfaces
preseed is rather simple minded, so when it comes to the network
configuration it can get tricky, especially in a multi-homed setup.

We convince it to comply by sending special boot options (see later)
But now we have one interface, configured via dhcp. That sucks. We want
to know the IPs, and we also want to configure the external interface
AND it should be the primary. Well, we do that in the late command, by
simply overwriting the configs.